Fix tests with Typescript
