+++ .prettierrc.js
@@ -1,4 +1,5 @@
 module.exports = {
   singleQuote: false,
+  endOfLine: "lf",
   trailingComma: "es5",
 };
