+++ package.json
@@ -9,7 +9,7 @@
     "format": "prettier --write {src,test}/**/*.ts",
     "test": "npm run lint && npm run test:only",
     "test:cover": "npm run test:only -- --coverage",
-    "test:prod": "cross-env BABEL_ENV=production npm run test",
+    "test:prod": "cross-env NODE_ENV=production npm run test",
     "test:only": "jest --verbose",
     "test:watch": "npm run test:only -- --watch",
     "lint": "eslint src test",
@@ -40,8 +40,8 @@
     "@babel/cli": "^7.10.3",
     "@babel/core": "^7.10.3",
     "@babel/preset-env": "^7.10.3",
-    "@babel/preset-typescript": "^7.10.4",
     "@babel/register": "^7.10.3",
+    "@types/jest": "^26.0.14",
     "@typescript-eslint/eslint-plugin": "^4.3.0",
     "@typescript-eslint/parser": "^4.3.0",
     "@vue/test-utils": "^1.0.3",
@@ -62,6 +62,7 @@
     "lint-staged": "^10.2.11",
     "prettier": "^2.0.5",
     "rimraf": "^3.0.2",
+    "ts-jest": "^26.4.1",
     "typescript": "^4.0.3",
     "vue": "^2.6.11",
     "vue-class-component": "^7.2.6",
