+++ src/index.ts
@@ -1,8 +1,4 @@
 import "./classComponentHooks";
 
 export { createForm } from "./createForm";
-export { FormLeaveGuardMixin } from "./formLeaveGuardMixin";
-export { default as FormGuardMixin } from "./FormGuardMixin";
-
-// NOTE: 'createForm' should be used instead when using TypeScript!
-export { FormCreateMixin } from "./FormCreateMixin";
+export { FormGuardMixin } from "./FormGuardMixin";
